backend/requirements.txt
# PyTorch CPU-only — install from the PyTorch CPU wheel index
--extra-index-url https://download.pytorch.org/whl/cpu
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
backoff==2.2.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
durationpy==0.10
fastapi==0.115.6
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
gotrue==2.12.4
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.2
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.63
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.2.11
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy-extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
openai
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==24.2
pillow==12.1.1
postgrest==0.19.3
posthog==7.9.12
propcache==0.4.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.2
pyjwt==2.12.1
pypdf==5.1.0
pypika==0.51.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.3
rank-bm25==0.2.2
realtime==2.5.3
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.48
starlette==0.41.3
storage3==0.11.3
strenum==0.4.15
supabase==2.11.0
supafunc==0.9.4
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.10.0+cpu
tqdm==4.67.3
transformers==4.41.2
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
urllib3==2.6.3
uvicorn==0.34.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.23.0
zipp==3.23.0
